+++ packages/webcomponents/src/components/terminal-orders-list/terminal-orders-list-filters.tsx
@@ -3,13 +3,7 @@ import { filterParams, propsParams, clearParams } from './terminal-orders-list-p
 import { StyledHost } from '../../ui-components';
 import { TerminalOrderStatus, TerminalOrderType } from '../../api';
 import { convertToLocal, convertToUTC } from '../../utils/utils';
-import {
-  terminalOrdersListFilterMenu,
-  orderStatusTerminalOrdersListFilterParam,
-  orderTypeTerminalOrdersListFilterParam,
-  createdAfterTerminalOrdersListFilterParam,
-  createdBeforeTerminalOrdersListFilterParam
-} from '../../styles/parts';
+import { filterMenu, filterParam } from '../../styles/parts';
 
 @Component({
   tag: 'justifi-terminal-orders-list-filters',
@@ -70,7 +64,7 @@ export class TerminalOrdersListFilters {
 
     return (
       <StyledHost>
-        <table-filters-menu params={filterMenuParams} clearParams={clearParams} part={terminalOrdersListFilterMenu}>
+        <table-filters-menu params={filterMenuParams} clearParams={clearParams} part={filterMenu}>
           <div class="grid-cols-2 gap-3 p-1">
             <div class="p-2">
               <form-control-select
@@ -80,7 +74,7 @@ export class TerminalOrdersListFilters {
                 inputHandler={this.setParamsOnChange}
                 defaultValue={this.orderStatus || filterParams.order_status || ''}
                 disabled={!!this.orderStatus}
-                part={orderStatusTerminalOrdersListFilterParam}
+                part={filterParam}
               />
             </div>
             <div class="p-2">
@@ -91,7 +85,7 @@ export class TerminalOrdersListFilters {
                 inputHandler={this.setParamsOnChange}
                 defaultValue={this.orderType || filterParams.order_type || ''}
                 disabled={!!this.orderType}
-                part={orderTypeTerminalOrdersListFilterParam}
+                part={filterParam}
               />
             </div>
             <div class="p-2">
@@ -105,7 +99,7 @@ export class TerminalOrdersListFilters {
                 }
                 showTime
                 disabled={!!this.createdAfter}
-                part={createdAfterTerminalOrdersListFilterParam}
+                part={filterParam}
               />
             </div>
             <div class="p-2">
@@ -119,7 +113,7 @@ export class TerminalOrdersListFilters {
                 }
                 showTime
                 disabled={!!this.createdBefore}
-                part={createdBeforeTerminalOrdersListFilterParam}
+                part={filterParam}
               />
             </div>
           </div>