+++ packages/webcomponents/src/components/terminals-list/terminals-list-filters.tsx
@@ -1,14 +1,10 @@
 import { Component, h, Prop } from '@stencil/core';
 import { debounce } from 'lodash';
+import { convertToLocal, convertToUTC } from '../../utils/utils';
 import { filterParams, propsParams, clearParams } from './terminals-list-params-state';
 import { StyledHost } from '../../ui-components';
 import { ITerminalStatus } from '../../api';
-import {
-  terminalsListFilterMenu,
-  terminalIdTerminalsListFilterParam,
-  terminalOrderIdTerminalsListFilterParam,
-  terminalStatusTerminalsListFilterParam
-} from '../../styles/parts';
+import { filterMenu, filterParam } from '../../styles/parts';
 
 @Component({
   tag: 'justifi-terminals-list-filters',
@@ -18,6 +14,8 @@ export class TerminalsListFilters {
   @Prop() terminalId?: string;
   @Prop() terminalOrderId?: string;
   @Prop() terminalStatus?: ITerminalStatus;
+  @Prop() createdAfter?: string;
+  @Prop() createdBefore?: string;
 
   private debouncedSetParamsOnChange: (name: string, value: string) => void;
 
@@ -28,7 +26,9 @@ export class TerminalsListFilters {
     const propsToSet = {
       terminal_id: this.terminalId,
       status: this.terminalStatus,
-      terminal_order_id: this.terminalOrderId
+      terminal_order_id: this.terminalOrderId,
+      created_after: this.createdAfter,
+      created_before: this.createdBefore
     };
 
     Object.entries(propsToSet).forEach(([key, value]) => {
@@ -41,6 +41,11 @@ export class TerminalsListFilters {
   setParamsOnChange = (name: string, value: string) => {
     filterParams[name] = value;
   };
+
+  handleDateInput = (name: string, value: string) => {
+    const utcDate = convertToUTC(value, { setExactTime: true });
+    this.setParamsOnChange(name, utcDate);
+  }
 
   get terminalStatusOptions(): { label: string, value: ITerminalStatus | '' }[] {
     return [
@@ -58,7 +63,7 @@ export class TerminalsListFilters {
 
     return (
       <StyledHost>
-        <table-filters-menu params={filterMenuParams} clearParams={clearParams} part={terminalsListFilterMenu}>
+        <table-filters-menu params={filterMenuParams} clearParams={clearParams} part={filterMenu}>
           <div class="grid-cols-2 gap-3 p-1">
             <div class="p-2">
               <form-control-text
@@ -67,7 +72,7 @@ export class TerminalsListFilters {
                 inputHandler={this.debouncedSetParamsOnChange}
                 defaultValue={this.terminalId || filterParams.terminal_id}
                 disabled={!!this.terminalId}
-                part={terminalIdTerminalsListFilterParam}
+                part={filterParam}
               />
             </div>
             <div class="p-2">
@@ -77,7 +82,7 @@ export class TerminalsListFilters {
                 inputHandler={this.debouncedSetParamsOnChange}
                 defaultValue={this.terminalOrderId || filterParams.terminal_order_id}
                 disabled={!!this.terminalOrderId}
-                part={terminalOrderIdTerminalsListFilterParam}
+                part={filterParam}
               />
             </div>
             <div class="p-2">
@@ -88,7 +93,33 @@ export class TerminalsListFilters {
                 inputHandler={this.setParamsOnChange}
                 defaultValue={this.terminalStatus || filterParams.status || ''}
                 disabled={!!this.terminalStatus}
-                part={terminalStatusTerminalsListFilterParam}
+                part={filterParam}
+              />
+            </div>
+            <div class="p-2">
+              <form-control-date
+                name="created_after"
+                label="Created After"
+                inputHandler={this.handleDateInput}
+                defaultValue={
+                  convertToLocal(this.createdAfter || filterParams.created_after, { showInputDateTime: true })
+                }
+                showTime
+                disabled={!!this.createdAfter}
+                part={filterParam}
+              />
+            </div>
+            <div class="p-2">
+              <form-control-date
+                name="created_before"
+                label="Created Before"
+                inputHandler={this.handleDateInput}
+                defaultValue={
+                  convertToLocal(this.createdBefore || filterParams.created_before, { showInputDateTime: true })
+                }
+                showTime
+                disabled={!!this.createdBefore}
+                part={filterParam}
               />
             </div>
           </div>