+++ apps/component-examples/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @repo/component-examples
 
+## 1.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @justifi/webcomponents@6.10.4
+
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @justifi/webcomponents@6.10.3
+
+## 1.2.0
+
+### Minor Changes
+
+- 074d118: Fix applePayCancelled event was not being emitted
+
+## 1.1.18
+
+### Patch Changes
+
+- Updated dependencies [bbb495f]
+  - @justifi/webcomponents@6.10.2
+
+## 1.1.17
+
+### Patch Changes
+
+- Updated dependencies [3406f47]
+- Updated dependencies [b37b70b]
+  - @justifi/webcomponents@6.10.1
+
+## 1.1.16
+
+### Patch Changes
+
+- Updated dependencies [d69a93c]
+- Updated dependencies [3ef304c]
+- Updated dependencies [8c2e3d1]
+- Updated dependencies [47b454f]
+- Updated dependencies [05fb036]
+- Updated dependencies [ca0e302]
+- Updated dependencies [2a5cbbe]
+  - @justifi/webcomponents@6.10.0
+
+## 1.1.15
+
+### Patch Changes
+
+- Updated dependencies [79c1fac]
+- Updated dependencies [aea215b]
+- Updated dependencies [aea215b]
+  - @justifi/webcomponents@6.9.0
+
 ## 1.1.14
 
 ### Patch Changes